+++ http/src/main/scala/com/paypal/stingray/http/resource/ResourceDriver.scala
@@ -7,8 +7,8 @@ import scala.util._
 import spray.routing.RequestContext
 import com.paypal.stingray.http.util.HttpUtil
 import akka.actor.{ActorRef, ActorRefFactory}
-import scala.concurrent.duration.Duration
-import com.paypal.stingray.http.resource.HttpResourceActor.ResourceContext
+import scala.concurrent.duration.FiniteDuration
+import com.paypal.stingray.http.resource.HttpResourceActor.{RequestParser, ResourceContext}
 
 /**
  * Implementation of a basic HTTP request handling pipeline.
@@ -30,34 +30,33 @@ object ResourceDriver {
    * @tparam ParsedRequest the request after parsing
    * @return the rewritten request execution
    */
-  final def serveWithRewrite[ParsedRequest](resourceActor: ResourceContext => AbstractResourceActor,
+  final def serveWithRewrite[ParsedRequest <: AnyRef](resourceActor: ResourceContext => AbstractResourceActor,
                                             mbResponseActor: Option[ActorRef] = None,
-                                            recvTimeout: Duration = HttpResourceActor.defaultRecvTimeout,
-                                            processRecvTimeout: Duration = HttpResourceActor.defaultProcessRecvTimeout)
+                                            recvTimeout: FiniteDuration = HttpResourceActor.defaultRecvTimeout,
+                                            processRecvTimeout: FiniteDuration = HttpResourceActor.defaultProcessRecvTimeout)
                                            (rewrite: RewriteFunction[ParsedRequest])
                                            (implicit actorRefFactory: ActorRefFactory): RequestContext => Unit = {
     ctx: RequestContext =>
       rewrite(ctx.request).map {
         case (request, parsed) =>
-          val serveFn = serve(resourceActor, r => Success(parsed), mbResponseActor, recvTimeout, processRecvTimeout)
+          val serveFn = serve(resourceActor, (r => Success(parsed): Try[AnyRef]): RequestParser, mbResponseActor, recvTimeout, processRecvTimeout)
           serveFn(ctx.copy(request = request))
       }.recover {
         case e: Exception =>
-          ctx.complete(HttpResponse(InternalServerError, HttpUtil.coerceError(Option(e.getMessage).getOrElse("").getBytes(charsetUtf8))))
+          ctx.complete(HttpResponse(InternalServerError, HttpUtil.toJsonErrorsMap(Option(e.getMessage).getOrElse(""))))
       }
   }
 
   /**
    * Run the request on this resource
    * @param resourceActor function for creating the actorRef which will process the request
-   * @tparam ParsedRequest the request after parsing
    * @return the request execution
    */
-  final def serve[ParsedRequest](resourceActor: ResourceContext => AbstractResourceActor,
-                                 requestParser: HttpResourceActor.RequestParser[ParsedRequest],
+  final def serve(resourceActor: ResourceContext => AbstractResourceActor,
+                                 requestParser: HttpResourceActor.RequestParser,
                                  mbResponseActor: Option[ActorRef] = None,
-                                 recvTimeout: Duration = HttpResourceActor.defaultRecvTimeout,
-                                 processRecvTimeout: Duration = HttpResourceActor.defaultProcessRecvTimeout)
+                                 recvTimeout: FiniteDuration = HttpResourceActor.defaultRecvTimeout,
+                                 processRecvTimeout: FiniteDuration = HttpResourceActor.defaultProcessRecvTimeout)
                                 (implicit actorRefFactory: ActorRefFactory): RequestContext => Unit = {
     { ctx: RequestContext =>
       val actor = actorRefFactory.actorOf(HttpResourceActor.props(resourceActor, ctx, requestParser, mbResponseActor, recvTimeout, processRecvTimeout))