+++ http/src/main/scala/com/paypal/stingray/http/util/HttpUtil.scala
@@ -141,7 +141,7 @@ object HttpUtil {
     // TODO: convert Manifest patterns to use TypeTag, ClassTag when Jackson implements that
     JsonUtil.toJson(t) match {
       case Success(j) => HttpEntity(ContentTypes.`application/json`, j)
-      case Failure(e) => coerceError(Option(e.getMessage).getOrElse(""))
+      case Failure(e) => toJsonErrorsMap(Option(e.getMessage).getOrElse(""))
     }
   }
 
@@ -152,17 +152,17 @@ object HttpUtil {
    * @param body the body
    * @return an HttpEntity containing an error JSON body
    */
-  def coerceError(body: Array[Byte]): HttpEntity = {
-    toJsonBody(Map("errors" -> List(new String(body, charsetUtf8))))
-  }
-
-  /**
-   * Used under the covers to force simple error strings into a JSON format
-   * @param body the body
-   * @return an HttpEntity containing an error JSON body
-   */
-  def coerceError(body: String): HttpEntity = {
+  def toJsonErrorsMap(body: String): HttpEntity = {
     toJsonBody(Map("errors" -> List(body)))
   }
 
+  /**
+   * Used to convert error objects into JSON format.
+   * @param body object to convert
+   * @tparam T type of object
+   * @return JSONified body
+   */
+  def toJsonErrors[T : Manifest](body: T): HttpEntity = {
+    toJsonBody(body)
+  }
 }