+++ http/src/test/scala/com/paypal/stingray/http/tests/resource/DummyResourceService.scala
@@ -2,15 +2,13 @@ package com.paypal.stingray.http.tests.resource
 
 import com.paypal.stingray.common.service.ServiceNameComponent
 import spray.routing.Directives._
-import com.paypal.stingray.http.resource.{AbstractResourceActor, HttpResourceActor, ResourceServiceComponent, ResourceDriver}
-import scala.concurrent.Future
-import spray.http.{HttpRequest, HttpResponse}
-import scala.util.{Try, Success}
+import com.paypal.stingray.http.resource.{HttpResourceActor, ResourceServiceComponent, ResourceDriver}
+import spray.http.HttpRequest
+import scala.util.Try
 import com.paypal.stingray.akka.actor.ActorSystemComponent
-import akka.actor.{ActorRef, Props}
 import com.paypal.stingray.http.tests.resource.DummyResource.GetRequest
-import com.paypal.stingray.common.trys._
 import com.paypal.stingray.http.resource.HttpResourceActor.ResourceContext
+import com.paypal.stingray.http.resource.ResourceDriver.RewriteFunction
 
 /**
  * A dummy resource service implementation for use with [[com.paypal.stingray.http.tests.resource.DummyResource]].
@@ -24,8 +22,12 @@ trait DummyResourceService
   /** This resource */
   val dummy: ResourceContext => DummyResource = new DummyResource(_)
 
-  val parseRequest: HttpResourceActor.RequestParser[GetRequest] = { _ : HttpRequest =>
+  val parseRequest: HttpResourceActor.RequestParser = { _ : HttpRequest =>
     Try (GetRequest("bar"))
+  }
+
+  val rewriteRequest: RewriteFunction[GetRequest] = { req: HttpRequest =>
+    Try((req,GetRequest("bar")))
   }
 
   /** The route for this resource */
@@ -33,6 +35,11 @@ trait DummyResourceService
     path("ping") {
       get {
         ResourceDriver.serve(dummy, parseRequest)
+      }
+    } ~
+    path("ping-rewrite") {
+      get {
+        ResourceDriver.serveWithRewrite(dummy)(rewriteRequest)
       }
     }
   }